@@ -10,24 +10,30 @@
     "build-storybook": "build-storybook"
   },
   "dependencies": {
-    "core-js": "^2.6.5",
-    "gsap": "^2.1.3",
-    "vue": "^2.6.10"
+    "core-js": "^3.8.3",
+    "gsap": "^3.11.4",
+    "vue": "^3.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.6.4",
-    "@storybook/addon-actions": "^5.2.5",
-    "@storybook/addon-links": "^5.2.5",
-    "@storybook/addons": "^5.2.5",
-    "@storybook/vue": "^5.2.5",
-    "@vue/cli-plugin-babel": "^3.0.0-rc.12",
+    "@storybook/addon-actions": "^6.5.15",
+    "@storybook/addon-essentials": "^6.5.15",
+    "@storybook/addon-interactions": "^6.5.15",
+    "@storybook/addon-links": "^6.5.15",
+    "@storybook/builder-webpack4": "^6.5.15",
+    "@storybook/manager-webpack4": "^6.5.15",
+    "@storybook/testing-library": "^0.0.13",
+    "@storybook/vue3": "^6.5.15",
+    "@vue/cli-plugin-babel": "~5.0.8",
     "@vue/cli-plugin-eslint": "^3.0.0-rc.12",
-    "@vue/cli-service": "^3.0.0-rc.12",
+    "@vue/cli-service": "~5.0.8",
+    "@vue/compiler-sfc": "^3.1.0",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.6",
     "babel-preset-vue": "^2.0.2",
     "eslint": "^5.16.0",
     "eslint-plugin-vue": "^5.0.0",
+    "vue-loader": "^16.8.3",
     "vue-template-compiler": "^2.6.10"
   },
   "eslintConfig": {
